Changed the boundary condition at the outlet
In the slab code developed by ChongShen Ng, Steven Chong, Rui Yang and Naoki Hori, the outlet
boundary condition is an radiative/advective/non-reflecting boundary condition. The idea is to
have a speed that is somehow related to the Courant number and determines the speed of 
the Courant waves. Any disturbance at the outlet which travels in waves slower than this speed 
gets advected out. They solve d(q)/d(t) + C d(q)/d(z) = 0 at the outlet.

We tried this boundary condition and many more including imposing a fixed velocity
at the outlet. This always caused a numerical instability in the form waves entering the domain.
We found through trial and error that this is an instability related to the diffusive terms
since the diffusive terms in Z and Y direction are handled explicitly as compared to the slab code
where they are handled implicitly. Through trial and error, we found that imposing the double derivative
normal to the outflow as zero solves this issue (i.e. d^2(q)/dz^2 = 0). In a way, the solution at outflow 
is linearly interpolated from the interior. 

We still do not fully understand why it solves the issue though. However, the outflow is always going 
to be aphysical in nature and we do not care too much about the physical accuracy of the solution 
close to the outlet for the expected application of this code.

There is a minor instability still creeping up the wall above the outlet. It is yet to be determined if this is simply an issue with the resolution or with the implementation of the B.C.s itself. I guess that it has to do with the sharpness of the outlet in the x-direction (vertical) and deciding which grid points overlap the outlet and which ones don't.

Other than this, the boundary conditions are now imposed in main.F90 after initialisation. The time marcher does require one additional halo update as compared to that of the base AFiD, but this is because the halo update does not update the y-borders of the pencils at the outlet (nz coordinate).  The unnecessary 2D arrays in "ventilation_arrays" have been discarded and the module itself has been renamed to "ibm_arrays". The unnecessary copy and paste outlet subroutines are also removed as halo update does not disturb the nodes (1:nx,xstart(2):xend(2),nz) when periodicity is set to False while calling decomp_2d_init.