Moved the body IBM into SolveImpEqnUpdate routines
For stability reasons, it is recommended to solve the IBM enforcing implicity inside the SolveImpEquUpdate subroutines. This does enhance the stability but does not fix the issue of velocities diverging and code blowing up for 768^3 grid. 
Minor change to the Global Quanitites subroutine to write the header on out_stats.out with underscore instead of space characters.