Removed very slow OutVisc calculation
The improved code now initializes a 2D fields outvscx, outvscy and outvscz and uses the fields for artificial viscosity to stabilize the outlet flow instead of computing the expensive OutVisc subroutine at each timestep/substep. This remedies the extremely significant performance hit. Now the speed is almost the same as that of AFiD 2.0. Thanks to Chris (https://github.com/chowland) for recommending the scorep profiler (installed on Snellius) for finding out the cause of the slower speed.